requirements.txt
#requirements.txt

fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
python-dotenv
livekit-agents
livekit-plugins-google
pandas
openpyxl
python-dateutil
pydantic-settings
fuzzywuzzy
python-levenshtein
holidays
redis
cryptography
celery
apscheduler
prometheus-client
structlog
twilio